+++ test/plural_rules/decimal_plural_rule_test.exs
@@ -6,8 +6,7 @@ defmodule Decimal.PluralRule.Test do
   @modules [TestBackend.Cldr.Number.Cardinal, TestBackend.Cldr.Number.Ordinal]
 
   for module <- @modules,
-      # module.known_locale_names,
-      locale_name <- ["kw"],
+      locale_name <- module.known_locale_names,
       {category, rule} <- module.plural_rules_for(locale_name),
       sample_type <- @sample_types,
       one_rule <- rule[sample_type] || [] do