+++ test/substitution_test.exs
@@ -8,3 +8,4 @@ defmodule Cldr.Substitution.Test do
     assert Cldr.Substitution.substitute("1", template) == [string]
   end
 end
+